requirements.txt
fastapi
uvicorn
python-dotenv
openai==1.55.3
openai
pydantic
